requirements.txt
aniso8601==2.0.0
asn1crypto==0.24.0
attrs==17.4.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
codecov==2.0.15
coverage==4.5.1
cryptography==2.1.4
flake8==3.5.0
flake8-isort==2.3
flake8-polyfill==1.0.2
Flask==0.12.2
Flask-RESTful==0.3.6
flask-tinyauth==0.0.2
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
pluggy==0.6.0
py==1.5.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.5.3
pytest==3.4.0
pytest-cov==2.5.1
pytz==2018.3
requests==2.18.4
six==1.11.0
testfixtures==5.4.0
urllib3==1.22
Werkzeug==0.14.1
